# backendserver/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ticketing API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://0b44d62428f0.ngrok-free.app",
      "description": "Development Server"
    }
  ],
  "paths": {
...
    "/auth/google/login": {
      "get": {
        "summary": "Google Login",
        "description": "Initiates the Google OAuth flow.",
        "operationId": "google_login_auth_google_login_get",
        "parameters": [
          {
            "name": "booking_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          },
          {
            "name": "redirect_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Redirect Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/login/whatsapp": {
      "post": {
        "summary": "Google Login Whatsapp",
        "description": "Generate Google Login Link (REQUIRED).\n\nGenerates a Google OAuth link.\n**YOU MUST SEND THIS LINK TO THE USER** via WhatsApp so they can authorize access.\nThis step is **REQUIRED** before you can send emails or create calendar events.",
        "operationId": "google_login_whatsapp_auth_google_login_whatsapp_post",
        "parameters": [
          {
            "name": "whatsapp_user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "WhatsApp user ID or phone number",
              "title": "Whatsapp User Id"
            },
            "description": "WhatsApp user ID or phone number"
          },
          {
            "name": "booking_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Optional booking ID",
              "title": "Booking Id"
            },
            "description": "Optional booking ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "summary": "Google Callback",
        "description": "Handles the callback from Google OAuth.\nUpdates OAuth session status and stores tokens.\nReturns a WhatsApp-friendly success page.",
        "operationId": "google_callback_auth_google_callback_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/status/{whatsapp_user_id}": {
      "get": {
        "summary": "Check Oauth Status",
        "description": "Webhook endpoint to check OAuth completion status for a WhatsApp user.\nReturns the latest OAuth session status.",
        "operationId": "check_oauth_status_auth_google_status__whatsapp_user_id__get",
        "parameters": [
          {
            "name": "whatsapp_user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Whatsapp User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/status": {
      "get": {
        "summary": "Check Oauth Status By State",
        "description": "Check OAuth status by state token.\nUseful for webhook callbacks.",
        "operationId": "check_oauth_status_by_state_auth_google_status_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "OAuth state token",
              "title": "State"
            },
            "description": "OAuth state token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}